Allow importing only broker types
