+++ chronicle/src/lib.rs
@@ -21,8 +21,7 @@ impl<T: Serialize> SocketMsg<T> {
                 todo!()
             }
             SocketMsg::Broker(v) => {
-                serde_json::to_string(&chronicle_broker::application::SocketMsg::ChronicleBroker(v))
-                    .map_err(|e| e.to_string())
+                serde_json::to_string(&chronicle_broker::BrokerSocketMsg::ChronicleBroker(v)).map_err(|e| e.to_string())
             }
             SocketMsg::Scylla(v) => {
                 serde_json::to_string(&scylla_rs::prelude::SocketMsg::Scylla(v)).map_err(|e| e.to_string())
